backend/pyproject.toml
[project]
name = "backend-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp",
    "pyventus",
    "supabase",
    "python-dotenv",
]